rename energy into stamina
